externalize code
