+++ examples/medplum-provider/src/pages/encounter/EncounterChart.tsx
@@ -4,223 +4,48 @@ import {
   ClinicalImpression,
   QuestionnaireResponse,
   Questionnaire,
-  Practitioner,
   Encounter,
   ChargeItem,
   Claim,
 } from '@medplum/fhirtypes';
 import { Loading, QuestionnaireForm, useMedplum } from '@medplum/react';
-import { Outlet, useLocation, useParams } from 'react-router';
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { Outlet, useParams } from 'react-router';
 import { showNotification } from '@mantine/notifications';
-import { createReference, deepEquals, getReferenceString, normalizeErrorString } from '@medplum/core';
+import { createReference, deepEquals, getReferenceString } from '@medplum/core';
 import { IconCircleOff } from '@tabler/icons-react';
 import { TaskPanel } from '../components/Task/TaskPanel';
 import { EncounterHeader } from '../components/Encounter/EncounterHeader';
 import { usePatient } from '../../hooks/usePatient';
-import { useEncounter } from '../../hooks/useEncounter';
 import { SAVE_TIMEOUT_MS } from '../../config/constants';
 import ChageItemPanel from '../components/ChargeItem/ChageItemPanel';
 import { VisitDetailsPanel } from '../components/Encounter/VisitDetailsPanel';
+import { useEncounterChart } from '../../hooks/useEncounterChart';
+import { useState, useCallback, useRef } from 'react';
+import { showErrorNotification } from '../../utils/notifications';
 
 export const EncounterChart = (): JSX.Element => {
   const { encounterId } = useParams();
   const medplum = useMedplum();
   const patient = usePatient();
-  const [encounter, setEncounter] = useState<Encounter | undefined>(useEncounter());
-  const [claim, setClaim] = useState<Claim | undefined>();
-  const location = useLocation();
-  const [practitioner, setPractitioner] = useState<Practitioner | undefined>();
-  const [tasks, setTasks] = useState<Task[]>([]);
-  const [clinicalImpression, setClinicalImpression] = useState<ClinicalImpression | undefined>();
-  const [questionnaireResponse, setQuestionnaireResponse] = useState<QuestionnaireResponse | undefined>();
-  const [chargeItems, setChargeItems] = useState<ChargeItem[]>([]);
   const saveTimeoutRef = useRef<ReturnType<typeof setTimeout>>();
   const [activeTab, setActiveTab] = useState<string>('notes');
+  const [isLoadingEncounter, setIsLoadingEncounter] = useState<boolean>(false);
 
-  useEffect(() => {
-    if (encounterId) {
-      medplum.readResource('Encounter', encounterId).then(setEncounter).catch(console.error);
-    }
-  }, [encounterId, medplum]);
-
-  const fetchTasks = useCallback(async (): Promise<void> => {
-    if (!encounter) {
-      return;
-    }
-    const taskResult = await medplum.searchResources('Task', `encounter=${getReferenceString(encounter)}`, {
-      cache: 'no-cache',
-    });
-
-    taskResult.sort((a: Task, b: Task) => {
-      const dateA = new Date(a.authoredOn || '').getTime();
-      const dateB = new Date(b.authoredOn || '').getTime();
-      return dateA - dateB;
-    });
-
-    setTasks(taskResult);
-  }, [medplum, encounter]);
-
-  const fetchClinicalImpressions = useCallback(async (): Promise<void> => {
-    if (!encounter) {
-      return;
-    }
-    const clinicalImpressionResult = await medplum.searchResources(
-      'ClinicalImpression',
-      `encounter=${getReferenceString(encounter)}`
-    );
-
-    const result = clinicalImpressionResult?.[0];
-    setClinicalImpression(result);
-
-    if (result?.supportingInfo?.[0]?.reference) {
-      const response = await medplum.readReference({ reference: result.supportingInfo[0].reference });
-      setQuestionnaireResponse(response as QuestionnaireResponse);
-    }
-  }, [medplum, encounter]);
-
-  const fetchChargeItems = useCallback(async (): Promise<void> => {
-    if (!encounter) {
-      return;
-    }
-    const chargeItems = await medplum.searchResources('ChargeItem', `context=${getReferenceString(encounter)}`);
-    setChargeItems(chargeItems);
-  }, [medplum, encounter]);
-
-  const fetchClaim = useCallback(async (): Promise<void> => {
-    if (!encounter) {
-      return;
-    }
-    const response = await medplum.searchResources('Claim', `encounter=${getReferenceString(encounter)}`);
-    setClaim(response[0]);
-  }, [encounter, medplum]);
-
-  useEffect(() => {
-    fetchTasks().catch((err) => {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-    fetchClinicalImpressions().catch((err) => {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-
-    fetchChargeItems().catch((err) => {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-
-    fetchClaim().catch((err) => {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-  }, [medplum, encounterId, fetchTasks, fetchClinicalImpressions, fetchChargeItems, fetchClaim, location.pathname]);
-
-  useEffect(() => {
-    const fetchPractitioner = async (): Promise<void> => {
-      if (encounter?.participant?.[0]?.individual) {
-        const practitionerResult = await medplum.readReference(encounter.participant[0].individual);
-        setPractitioner(practitionerResult as Practitioner);
-      }
-    };
-
-    fetchPractitioner().catch((err) => {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-  }, [encounter, medplum]);
-
-  const isUpdatingRef = useRef(false);
-
-  useEffect(() => {
-    if (isUpdatingRef.current) {
-      isUpdatingRef.current = false;
-      return;
-    }
-
-    const fetchChargeItemDefinitions = async (): Promise<void> => {
-      if (!chargeItems || chargeItems.length === 0) {
-        return;
-      }
-
-      const updatedItems = [...chargeItems];
-      let hasUpdates = false;
-
-      for (const [index, chargeItem] of chargeItems.entries()) {
-        if (chargeItem.definitionCanonical && chargeItem.definitionCanonical.length > 0) {
-          try {
-            const searchResult = await medplum.searchResources(
-              'ChargeItemDefinition',
-              `url=${chargeItem.definitionCanonical[0]}`
-            );
-            if (searchResult.length > 0) {
-              const chargeItemDefinition = searchResult[0];
-              try {
-                const applyResult = await medplum.post(
-                  medplum.fhirUrl('ChargeItemDefinition', chargeItemDefinition.id as string, '$apply'),
-                  {
-                    resourceType: 'Parameters',
-                    parameter: [
-                      {
-                        name: 'chargeItem',
-                        valueReference: {
-                          reference: getReferenceString(chargeItem),
-                        },
-                      },
-                    ],
-                  }
-                );
-
-                if (applyResult) {
-                  const updatedChargeItem = applyResult as ChargeItem;
-                  updatedItems[index] = updatedChargeItem;
-                  hasUpdates = true;
-                }
-              } catch (err) {
-                console.error('Error applying ChargeItemDefinition:', err);
-              }
-            }
-          } catch (err) {
-            console.error('Error fetching ChargeItemDefinition:', err);
-          }
-        }
-      }
-
-      if (hasUpdates) {
-        isUpdatingRef.current = true;
-        setChargeItems(updatedItems);
-      }
-    };
-
-    fetchChargeItemDefinitions().catch((err) => {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-  }, [chargeItems, medplum]);
+  const {
+    encounter,
+    claim,
+    practitioner,
+    tasks,
+    clinicalImpression,
+    questionnaireResponse,
+    chargeItems,
+    setEncounter,
+    setClaim,
+    setTasks,
+    setClinicalImpression,
+    setQuestionnaireResponse,
+    setChargeItems,
+  } = useEncounterChart(encounterId);
 
   const saveQuestionnaireResponse = async (response: QuestionnaireResponse): Promise<void> => {
     if (saveTimeoutRef.current) {
@@ -246,12 +71,7 @@ export const EncounterChart = (): JSX.Element => {
           setQuestionnaireResponse(updatedResponse);
         }
       } catch (err) {
-        showNotification({
-          color: 'red',
-          icon: <IconCircleOff />,
-          title: 'Error',
-          message: normalizeErrorString(err),
-        });
+        showErrorNotification(err);
       }
     }, SAVE_TIMEOUT_MS);
   };
@@ -260,12 +80,7 @@ export const EncounterChart = (): JSX.Element => {
     if (!questionnaireResponse) {
       const updatedResponse: QuestionnaireResponse = { ...response, status: 'in-progress' };
       saveQuestionnaireResponse(updatedResponse).catch((err) => {
-        showNotification({
-          color: 'red',
-          icon: <IconCircleOff />,
-          title: 'Error',
-          message: normalizeErrorString(err),
-        });
+        showErrorNotification(err);
       });
     } else {
       const equals = deepEquals(response.item, questionnaireResponse?.item);
@@ -276,12 +91,7 @@ export const EncounterChart = (): JSX.Element => {
           status: 'in-progress',
         };
         saveQuestionnaireResponse(updatedResponse).catch((err) => {
-          showNotification({
-            color: 'red',
-            icon: <IconCircleOff />,
-            title: 'Error',
-            message: normalizeErrorString(err),
-          });
+          showErrorNotification(err);
         });
       }
     }
@@ -291,7 +101,7 @@ export const EncounterChart = (): JSX.Element => {
     (updatedTask: Task): void => {
       setTasks(tasks.map((task) => (task.id === updatedTask.id ? updatedTask : task)));
     },
-    [tasks]
+    [tasks, setTasks]
   );
 
   const saveChargeItem = useCallback(
@@ -299,12 +109,7 @@ export const EncounterChart = (): JSX.Element => {
       try {
         return await medplum.updateResource(chargeItem);
       } catch (err) {
-        showNotification({
-          color: 'red',
-          icon: <IconCircleOff />,
-          title: 'Error',
-          message: normalizeErrorString(err),
-        });
+        showErrorNotification(err);
         return chargeItem;
       }
     },
@@ -322,7 +127,7 @@ export const EncounterChart = (): JSX.Element => {
         setChargeItems(chargeItems.map((item) => (item.id === savedChargeItem.id ? savedChargeItem : item)));
       }, SAVE_TIMEOUT_MS);
     },
-    [chargeItems, saveChargeItem]
+    [chargeItems, saveChargeItem, setChargeItems]
   );
 
   const handleEncounterStatusChange = useCallback(
@@ -352,15 +157,10 @@ export const EncounterChart = (): JSX.Element => {
         await medplum.updateResource(updatedEncounter);
         setEncounter(updatedEncounter);
       } catch (err) {
-        showNotification({
-          color: 'red',
-          icon: <IconCircleOff />,
-          title: 'Error',
-          message: normalizeErrorString(err),
-        });
+        showErrorNotification(err);
       }
     },
-    [encounter, medplum]
+    [encounter, medplum, setEncounter]
   );
 
   const handleTabChange = (tab: string): void => {
@@ -372,6 +172,7 @@ export const EncounterChart = (): JSX.Element => {
       return;
     }
 
+    setIsLoadingEncounter(true);
     if (saveTimeoutRef.current) {
       clearTimeout(saveTimeoutRef.current);
     }
@@ -381,12 +182,9 @@ export const EncounterChart = (): JSX.Element => {
         const savedEncounter = await medplum.updateResource(updatedEncounter);
         setEncounter(savedEncounter);
       } catch (err) {
-        showNotification({
-          color: 'red',
-          icon: <IconCircleOff />,
-          title: 'Error',
-          message: normalizeErrorString(err),
-        });
+        showErrorNotification(err);
+      } finally {
+        setIsLoadingEncounter(false);
       }
     }, SAVE_TIMEOUT_MS);
   };
@@ -435,20 +233,15 @@ export const EncounterChart = (): JSX.Element => {
       const createdClaim = await medplum.createResource(claim);
       setClaim(createdClaim);
     } catch (err) {
-      showNotification({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
+      showErrorNotification(err);
     }
-  }, [encounter, medplum, chargeItems, patient, practitioner]);
+  }, [encounter, medplum, chargeItems, patient, practitioner, setClaim]);
 
   const calculateTotalPrice = (chargeItems: ChargeItem[]): number => {
     return chargeItems.reduce((sum, item) => sum + (item.priceOverride?.value || 0), 0);
   };
 
-  if (!patient || !encounter || (clinicalImpression?.supportingInfo?.[0]?.reference && !questionnaireResponse)) {
+  if (!patient || !encounter) {
     return <Loading />;
   }
 
@@ -510,11 +303,7 @@ export const EncounterChart = (): JSX.Element => {
                       Total Calculated Price to Bill
                     </Text>
                     <Box>
-                      <TextInput
-                        w={300}
-                        value={`$${calculateTotalPrice(chargeItems)}`}
-                        readOnly
-                      />
+                      <TextInput w={300} value={`$${calculateTotalPrice(chargeItems)}`} readOnly />
                     </Box>
                   </Flex>
 
@@ -538,6 +327,7 @@ export const EncounterChart = (): JSX.Element => {
                     <Box mt="md">
                       <Button
                         fullWidth
+                        loading={isLoadingEncounter}
                         onClick={async () => {
                           await createClaimFromEncounter();
                         }}
